avancement p

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: ldermign <ldermign@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/12 09:05:57 by ldermign          #+#    #+#             */
-/*   Updated: 2021/02/11 13:40:29 by ldermign         ###   ########.fr       */
+/*   Updated: 2021/02/11 15:07:01 by ldermign         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -81,7 +81,7 @@ void			ft_putnbr_base_printf(long nbr, char *base, t_flag_len *len);
 char			*ft_strcat(char *dst, char *src);
 char			*ft_strncat(char *dst, char *src, size_t nb);
 void			ft_putnbr(unsigned int nbr, t_flag_len *len);
-void			ft_putnbr_adr(unsigned long nbr, t_flag_len *len);
+void			ft_putnbr_adr(unsigned int nbr, t_flag_len *len);
 char			*ft_strjoin(char const *s1, char const *s2);
 char			*ft_rev_string(char *str, int size);
